refactor plugin group
