+++ emulation/rocjitsu/plugins/execution_plugin.h
@@ -14,19 +14,9 @@
 #include "rocjitsu/isa/instruction.h"
 #include "rocjitsu/vm/amdgpu/wavefront.h"
 
-#include <cassert>
-#include <chrono>
-#include <cmath>
 #include <cstdint>
-#include <cstdio>
-#include <cstdlib>
-#include <iomanip>
-#include <iostream>
-#include <memory>
 #include <span>
 #include <string>
-#include <unordered_map>
-#include <vector>
 
 namespace rocjitsu {
 
@@ -127,293 +117,6 @@ private:
   friend class ExecutionPluginGroup;
   std::string name_;
   uint32_t slot_index_ = 0;
-};
-
-/// @brief Collection of plugins that delegates to each member.
-class ExecutionPluginGroup {
-  using Clock = std::chrono::steady_clock;
-
-  struct HookProfile {
-    uint64_t count = 0;
-    uint64_t last_sampled = 0;
-    uint64_t interval = 1;
-    double total_ns = 0;
-
-    double estimated_total_ms() const { return total_ns / 1e6; }
-  };
-
-public:
-  ExecutionPluginGroup()
-      : start_time_(Clock::now()), profiling_enabled_(std::getenv("RJ_HOOK_PROFILE") != nullptr) {}
-
-  bool add(std::unique_ptr<ExecutionPlugin> p) {
-    if (!p)
-      return false;
-    for (const auto &existing : plugins_)
-      if (existing->name() == p->name())
-        return false;
-    if (plugins_.empty()) {
-      active_start_ = &start_time_;
-      std::atexit([]() {
-        if (active_start_) {
-          double total = std::chrono::duration<double>(Clock::now() - *active_start_).count();
-          std::cerr << "[rocjitsu] total emulation time: " << std::fixed << std::setprecision(3)
-                    << total << " s" << std::endl;
-        }
-      });
-    }
-    p->slot_index_ = static_cast<uint32_t>(plugins_.size());
-    plugins_.push_back(std::move(p));
-    return true;
-  }
-
-  uint32_t num_plugins() const { return static_cast<uint32_t>(plugins_.size()); }
-
-  // -- Lifecycle --
-  void onInit() {
-    for (auto &p : plugins_)
-      p->onInit();
-  }
-
-  void onShutdown() {
-    for (auto &p : plugins_)
-      p->onShutdown();
-  }
-
-  // -- AMDGPU --
-  void beforeAmdgpuExecuteInstruction(uint64_t pc, const Instruction &inst, amdgpu::Wavefront &wf) {
-    if (profiling_enabled_) {
-      profiled_dispatch(prof_before_exec_, [&]() {
-        for (auto &p : plugins_)
-          p->beforeAmdgpuExecuteInstruction(pc, inst, wf);
-      });
-    } else {
-      for (auto &p : plugins_)
-        p->beforeAmdgpuExecuteInstruction(pc, inst, wf);
-    }
-  }
-
-  void afterAmdgpuExecuteInstruction(uint64_t pc, const Instruction &inst, amdgpu::Wavefront &wf) {
-    if (profiling_enabled_) {
-      profiled_dispatch(prof_after_exec_, [&]() {
-        for (auto &p : plugins_)
-          p->afterAmdgpuExecuteInstruction(pc, inst, wf);
-      });
-    } else {
-      for (auto &p : plugins_)
-        p->afterAmdgpuExecuteInstruction(pc, inst, wf);
-    }
-  }
-
-  void onAmdgpuRouteMemoryInstruction(const Instruction &inst, amdgpu::Wavefront &wf) {
-    if (profiling_enabled_) {
-      profiled_dispatch(prof_route_mem_, [&]() {
-        for (auto &p : plugins_)
-          p->onAmdgpuRouteMemoryInstruction(inst, wf);
-      });
-    } else {
-      for (auto &p : plugins_)
-        p->onAmdgpuRouteMemoryInstruction(inst, wf);
-    }
-  }
-
-  void onAmdgpuDispatchPacketProcessed(const KernelDispatchInfo &info) {
-    if (profiling_enabled_ && prof_before_exec_.count > 0)
-      print_profile_summary();
-    dispatch_kernel_names_[info.dispatch_id] = info.kernel_name;
-    current_kernel_name_ = info.kernel_name;
-    for (auto &p : plugins_)
-      p->onAmdgpuDispatchPacketProcessed(info);
-    reset_profiles();
-  }
-
-  void onAmdgpuDispatchExecutionBegin(uint32_t dispatch_id) {
-    for (auto &p : plugins_)
-      p->onAmdgpuDispatchExecutionBegin(dispatch_id);
-  }
-
-  void onAmdgpuDispatchExecutionEnd(uint32_t dispatch_id) {
-    if (profiling_enabled_ && prof_before_exec_.count > 0) {
-      auto it = dispatch_kernel_names_.find(dispatch_id);
-      if (it != dispatch_kernel_names_.end())
-        current_kernel_name_ = it->second;
-      print_profile_summary();
-    }
-    for (auto &p : plugins_)
-      p->onAmdgpuDispatchExecutionEnd(dispatch_id);
-    reset_profiles();
-  }
-
-  void onAmdgpuWorkgroupDispatched(uint32_t wg_id, uint32_t dispatch_id, uint32_t vgpr_count,
-                                   uint32_t sgpr_count, std::span<amdgpu::Wavefront *> wavefronts) {
-    if (profiling_enabled_) {
-      profiled_dispatch(prof_wg_dispatched_, [&]() {
-        for (auto &p : plugins_)
-          p->onAmdgpuWorkgroupDispatched(wg_id, dispatch_id, vgpr_count, sgpr_count, wavefronts);
-      });
-    } else {
-      for (auto &p : plugins_)
-        p->onAmdgpuWorkgroupDispatched(wg_id, dispatch_id, vgpr_count, sgpr_count, wavefronts);
-    }
-  }
-
-  void onAmdgpuWorkgroupCompleted(uint32_t dispatch_id, uint32_t wg_id) {
-    if (profiling_enabled_) {
-      profiled_dispatch(prof_wg_completed_, [&]() {
-        for (auto &p : plugins_)
-          p->onAmdgpuWorkgroupCompleted(dispatch_id, wg_id);
-      });
-    } else {
-      for (auto &p : plugins_)
-        p->onAmdgpuWorkgroupCompleted(dispatch_id, wg_id);
-    }
-  }
-
-  void onAmdgpuWavefrontDispatched(amdgpu::Wavefront &wf) {
-    if (profiling_enabled_) {
-      profiled_dispatch(prof_wf_dispatched_, [&]() {
-        for (auto &p : plugins_)
-          p->onAmdgpuWavefrontDispatched(wf);
-      });
-    } else {
-      for (auto &p : plugins_)
-        p->onAmdgpuWavefrontDispatched(wf);
-    }
-  }
-
-  void onAmdgpuWavefrontHalted(amdgpu::Wavefront &wf) {
-    if (profiling_enabled_) {
-      profiled_dispatch(prof_wf_halted_, [&]() {
-        for (auto &p : plugins_)
-          p->onAmdgpuWavefrontHalted(wf);
-      });
-    } else {
-      for (auto &p : plugins_)
-        p->onAmdgpuWavefrontHalted(wf);
-    }
-  }
-
-  void onAmdgpuReadVgpr(amdgpu::Wavefront *wf, uint32_t physical_reg, uint32_t lane,
-                        uint8_t byteMask = 0xF) {
-    if (profiling_enabled_) {
-      profiled_dispatch(prof_read_vgpr_, [&]() {
-        for (auto &p : plugins_)
-          p->onAmdgpuReadVgpr(wf, physical_reg, lane, byteMask);
-      });
-    } else {
-      for (auto &p : plugins_)
-        p->onAmdgpuReadVgpr(wf, physical_reg, lane, byteMask);
-    }
-  }
-
-  void onAmdgpuReadSgpr(amdgpu::Wavefront *wf, uint32_t physical_reg) {
-    if (profiling_enabled_) {
-      profiled_dispatch(prof_read_sgpr_, [&]() {
-        for (auto &p : plugins_)
-          p->onAmdgpuReadSgpr(wf, physical_reg);
-      });
-    } else {
-      for (auto &p : plugins_)
-        p->onAmdgpuReadSgpr(wf, physical_reg);
-    }
-  }
-
-  void onAmdgpuBarrierResolved(std::span<amdgpu::Wavefront *> wavefronts) {
-    if (profiling_enabled_) {
-      profiled_dispatch(prof_barrier_, [&]() {
-        for (auto &p : plugins_)
-          p->onAmdgpuBarrierResolved(wavefronts);
-      });
-    } else {
-      for (auto &p : plugins_)
-        p->onAmdgpuBarrierResolved(wavefronts);
-    }
-  }
-
-  // -- RISC-V --
-  void onRiscvExecuteInstruction(uint64_t pc, const Instruction &inst) {
-    for (auto &p : plugins_)
-      p->onRiscvExecuteInstruction(pc, inst);
-  }
-
-  bool empty() const { return plugins_.empty(); }
-
-  /// A shared empty group used as the default when no plugins are attached.
-  static std::shared_ptr<ExecutionPluginGroup> empty_group() {
-    static auto instance = std::make_shared<ExecutionPluginGroup>();
-    return instance;
-  }
-
-private:
-  // Absolute totals are inflated by ~18ns/call of Clock::now() overhead, which
-  // is extrapolated across all calls. Use (race - no-race) deltas instead.
-  template <typename F> void profiled_dispatch(HookProfile &prof, F &&fn) {
-    prof.count++;
-    if ((prof.count % prof.interval) == 0) {
-      auto t0 = Clock::now();
-      fn();
-      double ns = std::chrono::duration<double, std::nano>(Clock::now() - t0).count();
-      uint64_t gap = prof.count - prof.last_sampled;
-      prof.total_ns += ns * static_cast<double>(gap);
-      prof.last_sampled = prof.count;
-      prof.interval = static_cast<uint64_t>(std::sqrt(prof.count / 1000.0)) + 1;
-    } else {
-      fn();
-    }
-  }
-
-  void reset_profiles() {
-    prof_before_exec_ = {};
-    prof_after_exec_ = {};
-    prof_read_vgpr_ = {};
-    prof_read_sgpr_ = {};
-    prof_route_mem_ = {};
-    prof_barrier_ = {};
-    prof_wg_dispatched_ = {};
-    prof_wg_completed_ = {};
-    prof_wf_dispatched_ = {};
-    prof_wf_halted_ = {};
-  }
-
-  void print_profile_summary() const {
-    auto print_hook = [](const char *name, const HookProfile &p) {
-      if (p.count == 0)
-        return;
-      std::fprintf(stderr, "HOOK_PROFILE %-30s  calls=%-12lu  est_total=%.1f ms\n", name,
-                   static_cast<unsigned long>(p.count), p.estimated_total_ms());
-    };
-    const char *kname = current_kernel_name_.empty() ? "?" : current_kernel_name_.c_str();
-    std::fprintf(stderr, "HOOK_PROFILE --- %s ---\n", kname);
-    print_hook("beforeExecuteInstruction", prof_before_exec_);
-    print_hook("afterExecuteInstruction", prof_after_exec_);
-    print_hook("readVgpr", prof_read_vgpr_);
-    print_hook("readSgpr", prof_read_sgpr_);
-    print_hook("routeMemoryInstruction", prof_route_mem_);
-    print_hook("barrierResolved", prof_barrier_);
-    print_hook("workgroupDispatched", prof_wg_dispatched_);
-    print_hook("workgroupCompleted", prof_wg_completed_);
-    print_hook("wavefrontDispatched", prof_wf_dispatched_);
-    print_hook("wavefrontHalted", prof_wf_halted_);
-    std::fprintf(stderr, "HOOK_PROFILE ---\n");
-  }
-
-  std::vector<std::unique_ptr<ExecutionPlugin>> plugins_;
-  Clock::time_point start_time_;
-  static inline const Clock::time_point *active_start_ = nullptr;
-  bool profiling_enabled_;
-  std::string current_kernel_name_;
-  std::unordered_map<uint32_t, std::string> dispatch_kernel_names_;
-
-  HookProfile prof_before_exec_;
-  HookProfile prof_after_exec_;
-  HookProfile prof_read_vgpr_;
-  HookProfile prof_read_sgpr_;
-  HookProfile prof_route_mem_;
-  HookProfile prof_barrier_;
-  HookProfile prof_wg_dispatched_;
-  HookProfile prof_wg_completed_;
-  HookProfile prof_wf_dispatched_;
-  HookProfile prof_wf_halted_;
 };
 
 } // namespace rocjitsu
